# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Passagens",
...
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8044",
            "description": "Servidor API de Produção"
        }
    ],
    "paths": {
...
        "/api/classflights": {
            "get": {
                "tags": [
                    "ClassFlights"
                ],
                "summary": "Listar todas as classes de voos",
                "operationId": "27243e4dccba2f743f424f1acb740aea",
                "responses": {
                    "200": {
                        "description": "Lista de classes de voos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ClassFlights"
                ],
                "summary": "Criar uma nova classe de voo",
                "operationId": "c391be2a1164bfb3e4b51656acc242d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "flight_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "class_type": {
                                        "type": "string",
                                        "example": "Econômica"
                                    },
                                    "seat_quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Classe de voo criada com sucesso"
                    },
                    "400": {
                        "description": "Erro na criação da classe de voo"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/classflights/{id}": {
            "get": {
                "tags": [
                    "ClassFlights"
                ],
                "summary": "Mostrar uma classe de voo específica",
                "operationId": "56b70db01cd633a65be171d3f2f46930",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da classe de voo"
                    },
                    "404": {
                        "description": "Classe de voo não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ClassFlights"
                ],
                "summary": "Atualizar uma classe de voo",
                "operationId": "426f8960bdeaaf635796def59c10eb2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "flight_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "class_type": {
                                        "type": "string",
                                        "example": "Econômica"
                                    },
                                    "seat_quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Classe de voo atualizada com sucesso"
                    },
                    "404": {
                        "description": "Classe de voo não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ClassFlights"
                ],
                "summary": "Deletar uma classe de voo",
                "operationId": "17a640b6ba6ab31b714d1f986bcc2c75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Classe de voo deletada com sucesso"
                    },
                    "404": {
                        "description": "Classe de voo não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a Bearer token para autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}